refactor: replace kenat.test.js with dateconverter.test.js for Ethiopian calendar conversions
